Refactoring
Using Resource objects for the buildings and made a Totals resource to handle the current totals. Not sure if this is right or not but all classes that need to reference the current values can do so via that shared resource. Added logic to dynamically add building buttons to the screen and made a crude vertical scroller to house them.